# packages/specs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: todoer
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      # No bearerFormat: the token is a base64url payload with an HMAC
      # appended, not a JWT — there is no header or algorithm field. Claiming
      # bearerFormat: JWT here would send a client author to a JWT library
      # that cannot parse it.

  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }

    Health:
      type: object
      required: [status, version]
      properties:
        status: { type: string, enum: [ok] }
        version: { type: string }

    # Request schemas are closed (additionalProperties: false): the validator is the
    # runtime gate against drift. Response schemas stay open on purpose, so a server
    # that adds a field later does not break a client built against today's document.
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        # Bounds generous for a human, absurd for an attacker: 254 is
        # RFC 5321's own limit on an address, and a passphrase running past
        # a few hundred characters is a payload, not something to be typed.
        # Both endpoints are unauthenticated and the request body limit is
        # shared with /sync's 2mb batches, so the validator — which runs
        # ahead of scrypt — is what stands between an oversized password and
        # an ~80ms hash of it.
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 256 }

    SessionTokens:
      type: object
      required: [accessToken, accessExpiresAt, refreshToken]
      properties:
        accessToken: { type: string }
        accessExpiresAt: { type: string, format: date-time }
        # Sent in the body only; the web client's cookie transport (ADR 0011)
        # lands with the web client. Rotates on every use: store the newest.
        refreshToken: { type: string }

    # Passwords below keep LoginRequest's 8..256 bounds. Strength (letters,
    # digits and symbols) is a server rule, answered with 400, not a schema
    # one. Tokens and codes are capped at 512 so the validator, which runs
    # ahead of any hash or lookup, rejects a payload posing as one.
    RegisterRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 256 }
        invitation: { type: string, maxLength: 512 }

    RefreshRequest:
      type: object
      required: [refreshToken]
      additionalProperties: false
      properties:
        refreshToken: { type: string, maxLength: 512 }

    LogoutRequest:
      type: object
      additionalProperties: false
      properties:
        refreshToken: { type: string, maxLength: 512 }
        all: { type: boolean }

    PasswordChange:
      type: object
      required: [currentPassword, newPassword]
      additionalProperties: false
      properties:
        currentPassword: { type: string, minLength: 8, maxLength: 256 }
        newPassword: { type: string, minLength: 8, maxLength: 256 }

    InviteRequest:
      type: object
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 254 }

    Invitation:
      type: object
      required: [token, expiresAt]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }

    PasswordSet:
      type: object
      required: [password]
      additionalProperties: false
      properties:
        password: { type: string, minLength: 8, maxLength: 256 }

    ForgotRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email: { type: string, format: email, maxLength: 254 }

    ResetRequest:
      type: object
      required: [code, password]
      additionalProperties: false
      properties:
        code: { type: string, maxLength: 512 }
        password: { type: string, minLength: 8, maxLength: 256 }

    AccountDelete:
      type: object
      required: [password]
      additionalProperties: false
      properties:
        password: { type: string, minLength: 8, maxLength: 256 }

    Op:
      oneOf:
        - $ref: "#/components/schemas/OpCreate"
        - $ref: "#/components/schemas/OpSet"
        - $ref: "#/components/schemas/OpDelete"
      discriminator:
        propertyName: kind
        mapping:
          create: "#/components/schemas/OpCreate"
          set: "#/components/schemas/OpSet"
          delete: "#/components/schemas/OpDelete"

    OpCreate:
      type: object
      description: |
        Creates a row. Every id is a client-generated UUIDv7 (ADR 0005),
        except in `task_occurrence` and `task_tag`, whose id is derived:
        UUIDv5 in namespace 40e49f07-6ce6-46fc-b2de-65dd46253bf2 over
        `<taskId>:<occurrence>` (`<taskId>:` when occurrence is null) or
        `<taskId>:<tagId>`, ids lower-cased. The server recomputes it and
        rejects a mismatch. A create of an existing derived id merges its
        fields under per-field last-write-wins; rows of these two tables are
        never deleted, they are toggled with `set` (`state`, `attached`).
        Dates are `YYYY-MM-DD` strings. `@todoer/specs` exports the
        derivation (`taskOccurrenceId`, `taskTagId`) and the rrule parser.
      required: [opId, kind, table, id, fields, ts]
      additionalProperties: false
      properties:
        opId: { type: string, format: uuid }
        kind: { type: string, enum: [create] }
        table:
          {
            type: string,
            enum: [task, project, tag, task_tag, task_occurrence],
          }
        id: { type: string, format: uuid }
        fields: { type: object, additionalProperties: true }
        ts: { type: string, format: date-time }

    OpSet:
      type: object
      required: [opId, kind, table, id, field, value, ts]
      additionalProperties: false
      properties:
        opId: { type: string, format: uuid }
        kind: { type: string, enum: [set] }
        table:
          {
            type: string,
            enum: [task, project, tag, task_tag, task_occurrence],
          }
        id: { type: string, format: uuid }
        field: { type: string }
        value: {}
        ts: { type: string, format: date-time }
        # Optional: only destructive set targets (rrule) require it — see ADR 0004.
        baseVersion: { type: integer, minimum: 0 }

    OpDelete:
      type: object
      required: [opId, kind, table, id, baseVersion]
      additionalProperties: false
      properties:
        opId: { type: string, format: uuid }
        kind: { type: string, enum: [delete] }
        table:
          {
            type: string,
            enum: [task, project, tag, task_tag, task_occurrence],
          }
        id: { type: string, format: uuid }
        baseVersion: { type: integer, minimum: 0 }

    OpResult:
      type: object
      required: [opId, status]
      properties:
        opId: { type: string, format: uuid }
        status:
          type: string
          enum: [applied, duplicate, superseded, conflict, rejected]
        reason: { type: string }
        currentVersion: { type: integer }

    Change:
      type: object
      required: [table, id, seq, row]
      properties:
        table: { type: string }
        id: { type: string, format: uuid }
        seq: { type: integer }
        row: { type: object, additionalProperties: true }

    SyncRequest:
      type: object
      required: [since, ops]
      additionalProperties: false
      properties:
        # Bounded above at 2^53-1, not at a bigint's range: `seq` is a
        # Postgres bigint, but it travels in both directions as a JSON
        # number, and past 2^53-1 a JSON number is no longer the value that
        # was written — a client sending 9007199254740993 would be answered
        # with a cursor of 9007199254740992, lower than the one it sent.
        # Without the bound the server also had to answer a contract-legal
        # `1e300` with a 500, since it reaches Postgres as an out-of-range
        # bigint. `format: int64` is deliberately absent: it would widen the
        # bound this line exists to narrow, and it changes what the
        # generators emit for a field every client already reads as a number.
        since:
          type: integer
          minimum: 0
          maximum: 9007199254740991
          description: >-
            The cursor from the previous response. 0 asks for a snapshot:
            every live row and no tombstones, with a cursor that is never
            below the prune watermark. A request with since 0 is never
            answered 410; it is how a client recovers from one.
        ops:
          type: array
          items: { $ref: "#/components/schemas/Op" }
          maxItems: 1000

    SyncResponse:
      type: object
      required: [cursor, results, changes]
      properties:
        cursor: { type: integer }
        results:
          type: array
          items: { $ref: "#/components/schemas/OpResult" }
        changes:
          type: array
          items: { $ref: "#/components/schemas/Change" }

paths:
  /health:
    get:
      operationId: getHealth
      summary: Liveness and version
      security: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/login:
    post:
      operationId: postAuthLogin
      summary: Sign in
      description: >-
        Verifies the password and opens a new session. Every failure is the same
        401, whether the address is unknown or the password wrong. Repeated
        failures from one address or IP are answered 429 with Retry-After.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionTokens" }
        "401":
          description: bad credentials
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: too many failed attempts; see Retry-After
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/register:
    post:
      operationId: postAuthRegister
      summary: Create an account and sign in
      description: >-
        Open only while the instance has no users; the first account becomes the
        owner. Afterwards an invitation is required, single-use and, when it was
        issued for an address, valid only for that address.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: registered and signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionTokens" }
        "400":
          description: the password is too weak
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: registration is closed, or the invitation is invalid, spent, expired or issued for another address
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: that address is already registered
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: too many attempts; see Retry-After
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/refresh:
    post:
      operationId: postAuthRefresh
      summary: Exchange a refresh token for a new pair
      description: >-
        Spends the presented refresh token and returns a new access token and a
        new refresh token. A just-spent token presented again within about 30
        seconds returns the same refresh token with a freshly minted access
        token; a spent token past that window revokes the session. An expired, revoked or unknown token is 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: rotated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionTokens" }
        "401":
          description: invalid, expired or revoked refresh token
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: too many invalid refreshes from this IP — any refresh from it, valid or not, is refused until Retry-After
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/logout:
    post:
      operationId: postAuthLogout
      summary: Revoke the current session or all of them
      description: >-
        Revokes the session of the presented refresh token, or every session of
        the caller when `all` is true. Access tokens already issued live out their
        15 minutes. The body is required but may be `{}`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LogoutRequest" }
      responses:
        "204":
          description: signed out
        "401":
          description: missing or invalid bearer token
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/password:
    post:
      operationId: postAuthPassword
      summary: Change the caller's password
      description: >-
        Needs the current password. Revokes every session of the caller and
        returns a fresh pair, so the client that changed it stays signed in. The
        new password must be strong: letters, digits and symbols. Wrong current
        passwords are limited per user: too many are answered 429 with
        Retry-After.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChange" }
      responses:
        "200":
          description: password changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionTokens" }
        "400":
          description: the new password is too weak
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401":
          description: missing or invalid bearer token, or wrong current password
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: too many attempts; see Retry-After
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/invites:
    post:
      operationId: postAuthInvites
      summary: Issue an invitation
      description: >-
        Owner only. Returns a random single-use token valid for 7 days; the
        server keeps only its hash, so this response is the only time it is
        visible. With `email` the token registers only that address.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteRequest" }
      responses:
        "201":
          description: invitation issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "401":
          description: missing or invalid bearer token
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: the caller is not the owner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/users/{id}/password:
    post:
      operationId: postAuthUserPassword
      summary: Set another user's password
      description: >-
        Owner only. Sets the password directly (no reset code) and revokes
        every session of that user. The new password must be strong.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordSet" }
      responses:
        "204":
          description: password set
        "400":
          description: the password is too weak
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401":
          description: missing or invalid bearer token
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: the caller is not the owner, or targets themselves (use POST /auth/password)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: no such user
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/forgot:
    post:
      operationId: postAuthForgot
      summary: Ask for a password-reset mail
      description: >-
        Mail is not configured in this version, so this always answers 503.
        Recovery is the owner setting the password, or, for the owner, the host
        script that prints a one-time reset code for `/auth/reset`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotRequest" }
      responses:
        "503":
          description: mail is not configured
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/reset:
    post:
      operationId: postAuthReset
      summary: Set a password with a one-time reset code
      description: >-
        The code is single-use and valid for 15 minutes. Revokes every session of
        the user. An invalid, spent or expired code is the same 400 as a weak
        password.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetRequest" }
      responses:
        "204":
          description: password reset
        "400":
          description: the password is too weak, or the code is invalid
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: too many attempts; see Retry-After
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /auth/account:
    delete:
      operationId: deleteAuthAccount
      summary: Delete the caller's account
      description: >-
        Needs the current password. Purges the user and everything they own,
        without tombstones. The owner cannot delete the account while other users
        exist. Attempts per IP address are limited: too many are answered 429 with
        Retry-After.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountDelete" }
      responses:
        "204":
          description: account deleted
        "401":
          description: missing or invalid bearer token, or wrong password
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: the owner cannot be deleted while other users exist
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: too many attempts; see Retry-After
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  /sync:
    post:
      operationId: postSync
      summary: Push operations and pull changes
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SyncRequest" }
      responses:
        "200":
          description: synchronised
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncResponse" }
        "401":
          description: missing or invalid bearer token
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "410":
          description: >-
            The cursor is older than the prune watermark: tombstones it has
            not seen are gone. Discard the local replica and repeat with
            since 0. Operations in this request were still applied; resend
            them, and each replays its original outcome.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "413":
          description: batch larger than the server accepts
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: unexpected error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }